backend/pyproject.toml
[tool.poetry]
name = "admin-crm-platform"
version = "1.0.0"
description = "Admin Platform - Mini CRM for Sale & Marketing"
authors = ["Jacob <admin@company.vn>"]
readme = "README.md"
packages = [{include = "admin_crm", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
grpcio = "^1.62.0"
grpcio-tools = "^1.62.0"
grpcio-reflection = "^1.62.0"
protobuf = "^5.26.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.29"}
asyncpg = "^0.29.0"
alembic = "^1.13.1"
pydantic = "^2.7.0"
pydantic-settings = "^2.2.0"
argon2-cffi = "^23.1.0"
pyjwt = {extras = ["crypto"], version = "^2.8.0"}
python-dotenv = "^1.0.1"
structlog = "^24.1.0"
uvloop = {version = "^0.19.0", markers = "sys_platform != 'win32'"}

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.0"
pytest-asyncio = "^0.23.6"
pytest-cov = "^5.0.0"
grpcio-testing = "^1.62.0"
ruff = "^0.3.4"
mypy = "^1.9.0"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = true
plugins = ["pydantic.mypy"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
